# Simplify Map.Render()