--- a/render/render_test.go
+++ b/render/render_test.go
@@ -36,8 +36,8 @@ func TestReduceRender(t *testing.T) {
 func TestMapRender1(t *testing.T) {
 	// 1. Check when we return false, the node gets filtered out
 	mapper := render.Map{
-		MapFunc: func(nodes report.Node) report.Nodes {
-			return report.Nodes{}
+		MapFunc: func(nodes report.Node) report.Node {
+			return report.Node{}
 		},
 		Renderer: mockRenderer{Nodes: report.Nodes{
 			"foo": report.MakeNode("foo"),
@@ -53,10 +53,8 @@ func TestMapRender1(t *testing.T) {
 func TestMapRender2(t *testing.T) {
 	// 2. Check we can remap two nodes into one
 	mapper := render.Map{
-		MapFunc: func(nodes report.Node) report.Nodes {
-			return report.Nodes{
-				"bar": report.MakeNode("bar"),
-			}
+		MapFunc: func(nodes report.Node) report.Node {
+			return report.MakeNode("bar")
 		},
 		Renderer: mockRenderer{Nodes: report.Nodes{
 			"foo": report.MakeNode("foo"),
@@ -75,9 +73,9 @@ func TestMapRender2(t *testing.T) {
 func TestMapRender3(t *testing.T) {
 	// 3. Check we can remap adjacencies
 	mapper := render.Map{
-		MapFunc: func(nodes report.Node) report.Nodes {
+		MapFunc: func(nodes report.Node) report.Node {
 			id := "_" + nodes.ID
-			return report.Nodes{id: report.MakeNode(id)}
+			return report.MakeNode(id)
 		},
 		Renderer: mockRenderer{Nodes: report.Nodes{
 			"foo": report.MakeNode("foo").WithAdjacent("baz"),